### 2018-02-18 #585 @dwillcox
@zingale This is our JOSS pre review issue.
