### 2018-02-18 9868a74
Fix capitalization in bibtex.
