+++ paper/paper.bib
@@ -179,3 +179,33 @@ eprint = {https://doi.org/10.1137/0910062}
  url = {https://doi.org/10.7717/peerj-cs.103},
  doi = {10.7717/peerj-cs.103}
 }
+
+@ARTICLE{MAESTRO:Multilevel,
+   author = {A. Nonaka and A. S. Almgren and J. B. Bell and M. J. Lijewski and C. M. Malone and M. Zingale},
+    title = {{\tt MAESTRO}:An Adaptive Low Mach Number Hydrodynamics Algorithm for Stellar Flows},
+  journal = {The Astrophysical Journal Supplement Series},
+     year = {2010},
+   volume = {188},
+    pages = {358-383},
+     note = {paper V}
+}
+
+@ARTICLE{castro,
+   author = {{Almgren}, A.~S. and {Beckner}, V.~E. and {Bell}, J.~B. and
+             {Day}, M.~S. and {Howell}, L.~H. and {Joggerst}, C.~C. and {Lijewski}, M.~J. and
+             {Nonaka}, A. and {Singer}, M. and {Zingale}, M.},
+    title = "{CASTRO: A New Compressible Astrophysical Solver. I. Hydrodynamics and Self-gravity}",
+  journal = {The Astrophysical Journal},
+archivePrefix = "arXiv",
+   eprint = {1005.0114},
+ keywords = {equation of state, gravitation, hydrodynamics, methods:
+                  numerical, nuclear reactions, nucleosynthesis,
+                  abundances },
+     year = 2010,
+    month = jun,
+   volume = 715,
+    pages = {1221-1238},
+      doi = {10.1088/0004-637X/715/2/1221},
+   adsurl = {http://adsabs.harvard.edu/abs/2010ApJ...715.1221A},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
