### 2018-02-18 86862af
Remove Acknowledgements bold section heading
