### 2018-02-19 #588 @whedon
[REVIEW]: pynucastro: an interface to nuclear reaction rates and code generator for reaction network equations
**Submitting author:** @dwillcox (<a href="http://orcid.org/0000-0003-2300-5165">Donald Willcox</a>)
**Repository:** <a href="https://github.com/pynucastro/pynucastro" target ="_blank">https://github.com/pynucastro/pynucastro</a>
**Version:** 1.1.1
