### 2018-02-28 01ed2f8
this writes a net for sub-Ch
