### 2018-02-28 1408baa
Fix bug where capitalized nuclei were not located in Library.
