### 2018-03-01 f587990
Enforce consistently lowercase elements regardless the case of the abbreviation passed to Nucleus()
