### 2018-03-01 d36722c
Add Library.linking_nuclei taking a list of nuclei or nuclei abbreviations to link with rates. Returns a library object.
