### 2018-03-05 #588 @katyhuff
@jschwab you are the second reviewer and we are hoping you can contribute your particular expertise in nuclear astrophysics to this review. See below for instructions. (please note that first reviewer @kyleniemeyer has received his own set of instructions, and checklist, above).
