### 2018-03-05 c5d5b32
Fix links to example notebooks in README
