### 2018-03-05 pynucastro/pynucastro#31 @jschwab
Clarify requirements and testing command in README 
- [x] Explicitly state which Python versions are supported

- [x] Update Requirements section of README

+ The README says `scipy` is "only needed for the example integrator".  However,  one cannot successfully `import pynucastro` without this package.  Similarly, `sympy` and `ipywidgets` are unmentioned requirements.  Using the existing `requirements.txt` file in a clean venv does provide a working environment (though it pulls in more stuff than strictly needed, I think).

- [x] Clarify testing command

+ I believe the name of the py.test executable is more generically `py.test`.  (I understand some distributions do append the -3, but in a fresh venv this was not the case).
