### 2018-03-05 pynucastro/pynucastro#32 @jschwab
Example notebooks do not pass tests
When I run the tests for the example notebooks, 12 tests appear to fail.

I checked and I receive the same failure in `examples/pynucastro-examples.ipynb` when trying to evaluate this notebook "by hand".  (I didn't try the other examples yet.)

Here's the log.  For brevity, I excluded the tracebacks.  If you can't reproduce some or all of these failures, I can provide more information.

```
(venv) monolith:(master) pynucastro$ py.test --nbval examples --tb=no
============================================================ test session starts ============================================================
rootdir: /home/jschwab/Software/pynucastro, inifile:
plugins: cov-2.5.1, nbval-0.9.0
collected 49 items

examples/interactive-example.ipynb ....F.                                                                                             [ 12%]
examples/library-filtering-examples.ipynb ..F.FFFFFF.FF                                                                               [ 38%]
examples/pp-CNO-example.ipynb ....F.                                                                                                  [ 51%]
examples/pynucastro-examples.ipynb ..................F.....                                                                           [100%]

=================================================== 12 failed, 37 passed in 12.30 seconds ===================================================
```
