### 2018-03-05 pynucastro/pynucastro#33 @jschwab
Keep TODO list elsewhere
I think it would be better to keep the developer TODO list somewhere other than the README.

I doesn't seem like the currently existing list contains things the end user really needs to be aware of.
