### 2018-03-05 #588 @jschwab
Okay, just started the review.  I made it through the README and opened https://github.com/pynucastro/pynucastro/issues/30, https://github.com/pynucastro/pynucastro/issues/31, https://github.com/pynucastro/pynucastro/issues/32, https://github.com/pynucastro/pynucastro/issues/33.
