### 2018-03-05 #588 @dwillcox
Thanks Josiah! I'll make my way through these issues, I appreciate your careful feedback thus far and I look forward to your review.
