### 2018-03-05 a5a5854
Moved TODO from Readme into GitHub issues. closes #33
