+++ README.md
@@ -78,22 +78,3 @@ to test the notebooks, do:
 ```
 py.test-3 --nbval examples
 ```
-
-# TODO
-
-* output Cython code (rhs)
-
-  * probably want to use a cdef class for the TFactors:
-    http://stackoverflow.com/questions/21012348/cython-how-to-make-an-python-object-as-a-property-of-cython-class
-	http://cython.readthedocs.io/en/latest/src/tutorial/cdef_classes.html
-	http://cython.readthedocs.io/en/latest/src/userguide/early_binding_for_speed.html
-	
-
-
-
-* return a polynomial fit to a rate in a given interval
-
-* create a Ydot() class that can both output dYdt[nucleus] and
-  evaluate dYdt.
-  
-
