### 2018-03-05 ac1420c
Add sphinx_rtd_theme to requirements file.
