+++ requirements.txt
@@ -4,6 +4,7 @@ sympy>=1.1.1
 matplotlib==2.1.0
 jupyter>=1.0.0
 travis-sphinx
+sphinx_rtd_theme
 nbsphinx>=0.3.1
 pytest>=3.2.1
 nbval>=0.9.0
