### 2018-03-05 8c5ee79
Add dependencies to setup.py file.
