### 2018-03-05 4016e7d
Add scipy and ipywidgets to requirements file.
