### 2018-03-05 55ec30b
Revise wording for dependencies and unit test requirements.
