### 2018-03-05 bd8c727
Test older Python 3 versions with Travis.
