### 2018-03-05 62493b5
Fix typo in setup.py
