### 2018-03-05 4ad78e0
Clarify python version we need.
Earlier versions of Python 3 will work as well, but rates will enter
into the generated code in different orders because prior to Python
3.6, dictionaries were not guaranteed to maintain order. This causes
failures in the unit tests, which expect the dictionaries to preserve
order.
