+++ .travis.yml
@@ -2,11 +2,11 @@ language: python
 
 python:
   - "3.6"
-  - "3.5"
-  - "3.4"
-  - "3.3"
-  - "3.2"
-  - "3.1"
+# - "3.5"
+# - "3.4"
+# - "3.3"
+# - "3.2"
+# - "3.1"
 
 before_install:
   - export PATH=$(echo $PATH | tr ':' "\n" | sed '/\/opt\/python/d' | tr "\n" ":" | sed "s|::|:|g")
