+++ .travis.yml
@@ -2,7 +2,7 @@ language: python
 
 python:
   - "3.6"
-# - "3.5"
+  - "3.5"
 # - "3.4"
 # - "3.3"
 # - "3.2"
