### 2018-03-06 cd505f4
Replace dictionaries with OrderedDict for backwards compatibility.
