### 2018-03-06 5138963
Now that test pass with python 3.5, try 3.1-3.4 too.
