### 2018-03-06 4f014f2
Relax version requirement for scipy.
