### 2018-03-06 18ece7a
Test only python 3.4, 3.5, and 3.6 on Travis CI.
