### 2018-03-06 aff01bd
Clarify python version we support in the readme.
