### 2018-03-06 7aa4549
Require scipy>=1.0.0 in the requirements.txt.
