### 2018-03-06 pynucastro/pynucastro#31 @dwillcox
I've verified that the unit tests pass with Python 3.4, 3.5, and 3.6 and updated the Readme accordingly.

Travis CI will now test using those versions of Python as well.

I believe as of the following commit we satisfy the criteria outlined in this issue: https://github.com/pynucastro/pynucastro/commit/7aa45492294e5cd5ad2431f9bcfe09fcfaaf3731
