### 2018-03-06 pynucastro/pynucastro#31 @dwillcox
In addition, I've revised the dependency list in the Readme so one can start with a fresh python 3 virtual environment and install only the minimum packages listed in the Readme in order to import pynucastro successfully.

Readme now states that if the user wants to build the docs or run the test suite, they can install the packages provided by the requirements.txt file included.
