### 2018-03-06 pynucastro/pynucastro#31 @jschwab
Thanks!  I agree this is resolved.
