### 2018-03-06 3616f52
Add a template inputs file for the Microphysics burn_cell unit test with nuclei specified for clarity.
