+++ pynucastro/networks/starkiller_network.py
@@ -28,4 +28,4 @@ class StarKillerNetwork(BaseFortranNetwork):
         for i, n in enumerate(self.unique_nuclei):
             of.write("\n! {nuc: <5} initial mass fraction\n".format(nuc=str(n)))
             of.write("{}massfractions({}) = 0.0d0\n".format(
-                self.indent*n_indent, i))
+                self.indent*n_indent, i+1))
