### 2018-03-06 062b2bf
In readme, specify each notebook to run for unit testing, skip library filtering notebook.
