### 2018-03-06 6ac61d7
Use nbval-skip tags for the library filtering notebook instead of telling the user not to test it.
