### 2018-03-06 pynucastro/pynucastro#32 @dwillcox
Thanks very much for reporting these failures, I have tracked them down and corrected them as follows:

- `examples/interactive-example.ipynb`, `examples/pynucastro-examples.ipynb`, and `examples/pp-CNO-example.ipynb` generate a plot and Jupyter widget, which doesn't play nice with the unit test. I've tagged the relevant cells in those notebooks with `nbval-skip` so the test suite won't execute them.

- `examples/library-filtering-examples.ipynb` fails for a different reason. It reads a Reaclib library file, but we do not distribute a Reaclib library file with pynucastro, so it is not found and many errors follow. I've tagged all the cells in this notebook with `nbval-skip` so the test suite will not execute them either.

We have considered either getting permission to distribute a Reaclib library file or writing a separate library interface test program (see issue #29). This is something I intend to address in the near future.

For now, running `py.test --nbval examples` runs successfully for me:

```
$ py.test --nbval examples
========================================================= test session starts =========================================================
rootdir: /home/eugene/scratch/pynucastro, inifile:
plugins: ipynb-1.1.0, cov-2.5.1, nbval-0.9.0
collected 49 items

examples/interactive-example.ipynb ....s.                                                                                       [ 12%]
examples/library-filtering-examples.ipynb sssssssssssss                                                                         [ 38%]
examples/pp-CNO-example.ipynb ....s.                                                                                            [ 51%]
examples/pynucastro-examples.ipynb ...........s....s.s.....                                                                     [100%]

================================================ 31 passed, 18 skipped in 9.22 seconds ================================================
```
