### 2018-03-06 pynucastro/pynucastro#32 @jschwab
Great, I now see no failed tests  as well.

When I open `examples/pynucastro-examples.ipynb` in Jupyter and execute the notebook with "Run All", the cell

```
re = pyrl.Explorer(rc, comp)
re.explore()
```

fails with the Traceback

```
/usr/lib/python3.6/site-packages/ipywidgets/widgets/interaction.py in update(self, *args)
    215                     value = widget.get_interact_value()
    216                     self.kwargs[widget._kwarg] = value
--> 217                 self.result = self.f(**self.kwargs)
    218                 if self.auto_display and self.result is not None:
    219                     display(self.result)

~/Software/pynucastro/pynucastro/networks/rate_collection.py in _make_plot(self, logrho, logT)
    369
    370     def _make_plot(self, logrho, logT):
--> 371         self.rc.plot(rho=10.0**logrho, T=10.0**logT, comp=self.comp, size=self.size)
    372
    373     def explore(self):

~/Software/pynucastro/pynucastro/networks/rate_collection.py in plot(self, outfile, rho, T, comp, size, dpi)
    323         # which we can use for the colorbar
    324         if ydots is not None:
--> 325             plt.colorbar(edges_lc)
    326
    327         Zs = [n.Z for n in node_nuclei]

/usr/lib/python3.6/site-packages/matplotlib/pyplot.py in colorbar(mappable, cax, ax, **kw)
   2199         ax = gca()
   2200
-> 2201     ret = gcf().colorbar(mappable, cax = cax, ax=ax, **kw)
   2202     return ret
   2203 colorbar.__doc__ = matplotlib.colorbar.colorbar_doc

/usr/lib/python3.6/site-packages/matplotlib/figure.py in colorbar(self, mappable, cax, ax, use_gridspec, **kw)
   1862                 cax, kw = cbar.make_axes(ax, **kw)
   1863         cax._hold = True
-> 1864         cb = cbar.colorbar_factory(cax, mappable, **kw)
   1865
   1866         self.sca(current_ax)

/usr/lib/python3.6/site-packages/matplotlib/colorbar.py in colorbar_factory(cax, mappable, **kwargs)
   1366         cb = ColorbarPatch(cax, mappable, **kwargs)
   1367     else:
-> 1368         cb = Colorbar(cax, mappable, **kwargs)
   1369
   1370     cid = mappable.callbacksSM.connect('changed', cb.on_mappable_changed)

/usr/lib/python3.6/site-packages/matplotlib/colorbar.py in __init__(self, ax, mappable, **kw)
    919         # Ensure the given mappable's norm has appropriate vmin and vmax set
    920         # even if mappable.draw has not yet been called.
--> 921         mappable.autoscale_None()
    922
    923         self.mappable = mappable

AttributeError: 'list' object has no attribute 'autoscale_None'
```

I get the same error in the analogous cell of `examples/interactive-example.ipynb`.

I confirm that I did
```
monolith: ~$ jupyter nbextension enable --py --user widgetsnbextension
Enabling notebook extension jupyter-js-widgets/extension...
      - Validating: OK
````

Let me know if you need additional info.
