### 2018-03-06 pynucastro/pynucastro#32 @zingale
do you know what version of Jupyter you are using?  I seem to recall some of this widget stuff changed in 5.1.  I have it working in jupyter 5.4 after running that command.
