### 2018-03-06 pynucastro/pynucastro#32 @jschwab
```
monolith: ~$ jupyter --version
4.4.0
```
```
monolith: ~$ jupyter-notebook --version
5.3.0
```

I'll check with 5.4 and report back.
