### 2018-03-06 pynucastro/pynucastro#32 @zingale
I doubt it's an issue between 5.3 and 5.4.  Cause I did this functionality a while ago.  I wonder if it's a matplotlib backend thing?
