### 2018-03-06 pynucastro/pynucastro#32 @jschwab
I just tried again in a fresh venv after `pip install -r requirements.txt` and see this error.  (So that's using notebook 5.4.0 and matplotlib 2.1.0.)

I get the same error if I switch from `%matplotlib inline` to `%matplotlib notebook` at the start of the notebook.
