### 2018-03-06 pynucastro/pynucastro#32 @zingale
okay.  Let me google around some.  Unfortunately I can't reproduce the failure.  Maybe the list of edges from networkx is coming back empty somehow on your machine.
