### 2018-03-06 pynucastro/pynucastro#32 @jschwab
Thanks.  It draws most of the figure before failing.

![screenshot-on-2018-03-06-at-08 39 51](https://user-images.githubusercontent.com/772632/37045265-6400b08e-211a-11e8-934e-d4f91236b174.png)
