### 2018-03-06 pynucastro/pynucastro#32 @jschwab
Here's a pdb session with `locals()` at a couple key frames.  `colorbar` is not satisfied with a list of Patches.  Is colorbar happy with the same args for you?  Or is it getting something different?

```
> /usr/lib/python3.6/site-packages/matplotlib/colorbar.py(921)__init__()
-> mappable.autoscale_None()
(Pdb) locals()
{'kw': {'orientation': 'vertical', 'ticklocation': 'auto'}, 'mappable': [<matplotlib.patches.FancyArrowPatch object at 0x7f0634048da0>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050198>, <matplotlib.patches.FancyArrowPatch object at 0x7f06340504a8>, <matplotlib.patches.FancyArrowPatch object at 0x7f063403d978>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050a58>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050d68>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050fd0>, <matplotlib.patches.FancyArrowPatch object at 0x7f06340593c8>], 'ax': <matplotlib.axes._subplots.AxesSubplot object at 0x7f0634059710>, 'self': <matplotlib.colorbar.Colorbar object at 0x7f063400b7f0>}
(Pdb) u
> /usr/lib/python3.6/site-packages/matplotlib/colorbar.py(1368)colorbar_factory()
-> cb = Colorbar(cax, mappable, **kwargs)
(Pdb) u
> /usr/lib/python3.6/site-packages/matplotlib/figure.py(1864)colorbar()
-> cb = cbar.colorbar_factory(cax, mappable, **kw)
(Pdb) u
> /usr/lib/python3.6/site-packages/matplotlib/pyplot.py(2201)colorbar()
-> ret = gcf().colorbar(mappable, cax = cax, ax=ax, **kw)
(Pdb) u
> /home/jschwab/Software/pynucastro/pynucastro/networks/rate_collection.py(325)plot()
-> plt.colorbar(edges_lc)
(Pdb) locals()
{'edges_lc': [<matplotlib.patches.FancyArrowPatch object at 0x7f0634048da0>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050198>, <matplotlib.patches.FancyArrowPatch object at 0x7f06340504a8>, <matplotlib.patches.FancyArrowPatch object at 0x7f063403d978>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050a58>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050d68>, <matplotlib.patches.FancyArrowPatch object at 0x7f0634050fd0>, <matplotlib.patches.FancyArrowPatch object at 0x7f06340593c8>], 'weights': (-4.3582746639844085, -3.8879358773694093, -6.593885523771285, -5.314100503164024, -6.00816689231024, -1.057899875102397, -5.698173988887475, -5.965691950921178), 'edges': ((c12, n13, 0), (c13, n14, 0), (n13, c13, 0), (n13, o14, 0), (n14, o15, 0), (n15, c12, 0), (o14, n14, 0), (o15, n15, 0)), 'rate_weight': -5.965691950921178, 'p': n15, 'ydots': OrderedDict([(c12 + p --> n13, 4.3825344233265815e-05), (c13 + p --> n14, 0.00012943869407433355), (n13 --> c13, 2.5475016632596765e-07), (n13 + p --> o14, 4.8517620910445875e-06), (n14 + p --> o15, 9.813707457231496e-07), (n15 + p --> he4 + c12, 0.08751855225765924), (o14 --> n14, 2.0036691481625654e-06), (o15 --> n15, 1.0822012944765837e-06)]), 'r': o15 --> n15, 'n': o15, 'node_nuclei': [c12, c13, n13, n14, n15, o14, o15], 'G': <networkx.classes.multidigraph.MultiDiGraph object at 0x7f06343a0cc0>, 'dpi': 100, 'size': (800, 600), 'comp': <pynucastro.networks.rate_collection.Composition object at 0x7f06343a0240>, 'T': 100000000.0, 'rho': 10000.0, 'outfile': None, 'self': c12 + p --> n13
c13 + p --> n14
n13 --> c13
n13 + p --> o14
n14 + p --> o15
n15 + p --> he4 + c12
o14 --> n14
o15 --> n15
}
```
