### 2018-03-06 pynucastro/pynucastro#32 @zingale
for me, edges_lc is `<matplotlib.collections.LineCollection object at 0x7f593a8d4160>`
