### 2018-03-06 pynucastro/pynucastro#32 @zingale
oh... by any chance are you running networkx 2.2?

it looks like they changed the return type
