### 2018-03-06 pynucastro/pynucastro#32 @jschwab
Yes, I was running networkx 2.1.  I just changed the `requirements.txt` to ask for 2.0 and then things work great.
