### 2018-03-06 pynucastro/pynucastro#32 @zingale
huh... I hate it when return types change.

If looks like when arrows are drawn, then it is now a list of FancyArrow stuff instead of a LineCollection object.
