### 2018-03-06 #588 @jschwab
I made it through the "Software paper" and that part looks good to me (fixed a few typos in https://github.com/pynucastro/pynucastro/pull/37).

There's no DOI for [Zingale et al. 2017](http://adsabs.harvard.edu/abs/2017arXiv171106203Z) but it is on arXiv (and I'm guessing it will get a DOI in the future?).  The BibTeX entry and bib style mean that it doesn't get a hyperlink.  For now, perhaps just change adsurl -> url for that entry in the .bib to give it something clickable?
