### 2018-03-06 e451c66
some work on getting a colorbar working with NetworkX >= 2.1 not there yet
