### 2018-03-06 6cd3f53
Updated rate label format so its more human-readable. Fixes #40.
