### 2018-03-06 a0ef2c5
Set rate file path from the Library class interface. Fixes #39.
