### 2018-03-06 ec0c8e3
Require networkx 2.0 for now.
