+++ requirements.txt
@@ -1,4 +1,4 @@
-networkx>=2.0
+networkx==2.0
 scipy>=1.0.0
 numpy>=1.13.3
 sympy>=1.1.1
