### 2018-03-06 436becd
Test the notebooks on travis CI.
