### 2018-03-06 pynucastro/pynucastro#32 @dwillcox
Mike will revisit the networkx issue, but as a followup, I've changed requirements.txt to ask for version 2.0 of networkx until then.

With the recent changes to the notebook unit testing, the notebook tests now run on Travis CI along with the plain python tests. (https://github.com/pynucastro/pynucastro/commit/436becd936f4d3e1b15417ebe719313d700b27f3)
