### 2018-03-06 pynucastro/pynucastro#32 @jschwab
Sounds good.
