### 2018-03-07 99d5294
use the new networkx 2.1 style arrows
