### 2018-03-08 4be119e
Added conversion scripts to reformat partition function tables from the following sources into a standard form for pynucastro:
Thomas Rauscher and Friedrich-Karl Thielemann, Atomic Data and Nuclear
Data Tables, 75:1–351, 2000

Thomas Rauscher, Astrophysical Journal Supplement Series, 147:403-408,
2003

For Rauscher 2003, the conversion script handles both Table 2 (with
Finite Range Droplet Model inputs) and Table 3 (with inputs from the
extended Thomas-Fermi mass formula with shell quenching effects).
