### 2018-03-08 13ea76c
Add class for storing a partition function and interpolation.
