### 2018-03-08 99d938c
Added classes for storing partition functions, reading table files, and storing collections of partition function tables.
