### 2018-03-08 d28d229
Moved PartitionFunction into partition_functions.py
