### 2018-03-08 2abc979
Add _partition_function as Nucleus class variable. Add set and get methods.
