### 2018-03-08 e86ef63
In RateCollection, set the partition functions for unique nuclei.
