### 2018-03-08 559419a
Add PartitionFunction, PartitionFunctionCollection to imports.
