+++ pynucastro/nucdata/__init__.py
@@ -6,3 +6,4 @@ particular, the binding energies)
 from .binding_nuclide import BindingNuclide
 from .binding_table import BindingTable
 from .elements import Element, UnidentifiedElement, PeriodicTable
+from .partition_functions import PartitionFunction, PartitionFunctionCollection
