### 2018-03-08 409f760
Move binding energy table files into AtomicMassEvaluation subdirectory.
