### 2018-03-08 6798808
Consolidate binding energy classes into one file.
