### 2018-03-08 e9b173a
Rename binding energy table file for clarity.
