### 2018-03-08 7eaae63
Set partition functions for all the Nucleus objects contained in a RateCollection.
