### 2018-03-09 79737bf
Revised interface to PartitionFunctionCollection from Nucleus, Rate, RateCollection. Fixed minor bugs.
