### 2018-03-09 bf8bca6
Revise name of get method for PartitionFunctionCollection. Added call method for PartitionFunction.
