### 2018-03-09 abf31fd
Python networks can use partition functions.
