+++ examples/CNO_with_reverse/network_module.py
@@ -1,0 +1,375 @@
+import numpy as np
+from pynucastro.rates import Tfactors
+from pynucastro.nucdata import PartitionFunction, PartitionFunctionCollection
+
+ip = 0
+ihe4 = 1
+ic12 = 2
+ic13 = 3
+in13 = 4
+in14 = 5
+in15 = 6
+io14 = 7
+io15 = 8
+nnuc = 9
+
+A = np.zeros((nnuc), dtype=np.int32)
+
+A[ip] = 1
+A[ihe4] = 4
+A[ic12] = 12
+A[ic13] = 13
+A[in13] = 13
+A[in14] = 14
+A[in15] = 15
+A[io14] = 14
+A[io15] = 15
+
+species_names = ["p", "he4", "c12", "c13", "n13", "n14", "n15", "o14", "o15"]
+
+partfun_collection = PartitionFunctionCollection()
+
+pf_p = partfun_collection.get_partition_function('p')
+pf_he4 = partfun_collection.get_partition_function('he4')
+pf_c12 = partfun_collection.get_partition_function('c12')
+pf_c13 = partfun_collection.get_partition_function('c13')
+pf_n13 = partfun_collection.get_partition_function('n13')
+pf_n14 = partfun_collection.get_partition_function('n14')
+pf_n15 = partfun_collection.get_partition_function('n15')
+pf_o14 = partfun_collection.get_partition_function('o14')
+pf_o15 = partfun_collection.get_partition_function('o15')
+
+def n13__c13(tf):
+    # n13 --> c13
+    rate = 0.0
+    
+    # wc12w
+    rate += np.exp(  -6.7601)
+    
+    return rate
+
+def o14__n14(tf):
+    # o14 --> n14
+    rate = 0.0
+    
+    # wc12w
+    rate += np.exp(  -4.62354)
+    
+    return rate
+
+def o15__n15(tf):
+    # o15 --> n15
+    rate = 0.0
+    
+    # wc12w
+    rate += np.exp(  -5.17053)
+    
+    return rate
+
+def n13__p_c12(tf):
+    # n13 --> p + c12
+    rate = 0.0
+    
+    # ls09n
+    rate += np.exp(  40.0408 + -22.5475*tf.T9i + -13.692*tf.T913i + -0.230881*tf.T913
+                  + 4.44362*tf.T9 + -3.15898*tf.T953 + 0.833333*tf.lnT9)
+    # ls09r
+    rate += np.exp(  40.4354 + -26.326*tf.T9i + -5.10735*tf.T913i + -2.24111*tf.T913
+                  + 0.148883*tf.T9)
+    
+    rate = rate * pf_p(tf.T9*1.0e9)
+    rate = rate * pf_c12(tf.T9*1.0e9)
+    rate = rate / pf_n13(tf.T9*1.0e9)
+    
+    return rate
+
+def n14__p_c13(tf):
+    # n14 --> p + c13
+    rate = 0.0
+    
+    # nacrr
+    rate += np.exp(  37.1528 + -93.4071*tf.T9i + -0.196703*tf.T913
+                  + 0.142126*tf.T9 + -0.0238912*tf.T953)
+    # nacrr
+    rate += np.exp(  38.3716 + -101.18*tf.T9i)
+    # nacrn
[…]
