### 2018-03-09 5c88b21
Switch to interpolation in GK and log-space for temperature and partition functions, respectively. This is necessary to ensure the spline interpolation is smooth for the high temperature partition functions due to large variation in order of magnitude of the partition functions.
Add supporting routines for iterating through PartitionFunctionCollection.
