+++ pynucastro/nucdata/partition_functions.py
@@ -27,7 +27,10 @@ class PartitionFunction(object):
     
     Units of temperature are Kelvins.
     """
-    def __init__(self, name=None, temperature=None, partition_function=None):
+    def __init__(self, nucleus=None, name=None, temperature=None, partition_function=None):
+        if type(nucleus) == str:
+            nucleus = Nucleus(nucleus)
+        self.nucleus = nucleus
         self.name = name
         self.temperature = temperature
         self.partition_function = partition_function
@@ -38,9 +41,9 @@ class PartitionFunction(object):
             len(self.temperature) == len(self.partition_function)):
             self.construct_spline_interpolant()
         else:
-            # For a null partition function, return 1.0 when evaluated.
+            # For a null partition function, return log(pf) = 0.0 when evaluated.
             self.interpolant_order = 0
-            self.interpolant = lambda x: 1.0
+            self.interpolant = lambda x: 0.0
 
     def lower_partition(self):
         return self.partition_function[0]
@@ -67,6 +70,7 @@ class PartitionFunction(object):
         interpolant for the returned PartitionFunction of order equal
         to the maximum order of the added PartitionFunction objects.
         """
+        assert(self.nucleus == other.nucleus)
         assert(self.upper_temperature() < other.lower_temperature() or
                self.lower_temperature() > other.upper_temperature())
         if self.upper_temperature() < other.lower_temperature():
@@ -80,9 +84,8 @@ class PartitionFunction(object):
         partition_function = np.array(list(lower.partition_function) +
                                       list(upper.partition_function))
         name = '{} + {}'.format(lower.name, upper.name)
-        newpf = PartitionFunction(name=name, temperature=temperature,
+        newpf = PartitionFunction(nucleus=self.nucleus, name=name, temperature=temperature,
                                   partition_function=partition_function)
-
         if self.interpolant_order and other.interpolant_order:
             order = max(self.interpolant_order, other.interpolant_order)
         elif self.interpolant_order:
@@ -101,9 +104,12 @@ class PartitionFunction(object):
         Construct an interpolating univariate spline of order >= 1 and
         order <= 5 using the scipy InterpolatedUnivariateSpline
         implementation. 
+
+        Interpolate in log space for the partition function and in GK
+        for temperature.
         """
-        self.interpolant = InterpolatedUnivariateSpline(self.temperature,
-                                                        self.partition_function,
+        self.interpolant = InterpolatedUnivariateSpline(self.temperature/1.0e9,
+                                                        np.log10(self.partition_function),
                                                         k=order)
         self.interpolant_order = order
 
@@ -111,17 +117,19 @@ class PartitionFunction(object):
         """
         If the interpolant is already constructed, then evaluate it at
         temperature T. Otherwise, generate an exception.
+
+        Convert T to GK for evaluation. Return un-log-scaled partition function.
         """
         assert(self.interpolant)
         try:
-            T = float(T)
+            T = float(T)/1.0e9
         except:
             raise
         else:
             if self.interpolant_order == 0:
-                return self.interpolant(T)
+                return 10.0**self.interpolant(T)
             else:
-                return self.interpolant(T, ext='const')
+                return 10.0**self.interpolant(T, ext='const')
 
 class PartitionFunctionTable(object):
     """ 
@@ -143,6 +151,15 @@ class PartitionFunctionTable(object):
         """
         assert(not nuc in self._partition_functions)
         self._partition_functions[str(nuc)] = pfun
+
+    def get_nuclei(self):
+        """
+        Return list of Nucleus objects assigned partition functions in this table.
+        """
[…]
