+++ pynucastro/networks/tests/test_python_net.py
@@ -1,11 +1,6 @@
 # unit tests for rates
-import math
-import os
-
 import pynucastro.networks as networks
 import pynucastro.rates as rates
-
-from pytest import approx
 
 class TestPythonNetwork(object):
     @classmethod
