### 2018-03-10 dc91ede
some flake8 fixes
