+++ pynucastro/networks/tests/test_python_net.py
@@ -2,6 +2,7 @@
 import pynucastro.networks as networks
 import pynucastro.rates as rates
 
+
 class TestPythonNetwork(object):
     @classmethod
     def setup_class(cls):
@@ -35,13 +36,11 @@ class TestPythonNetwork(object):
         assert ydot == "rho*Y[ic13]*Y[ip]*lambda_p_c13__n14" or \
                ydot == "rho*Y[ip]*Y[ic13]*lambda_p_c13__n14"
 
-
     def test_jacobian_string(self):
         jac = self.pyn.jacobian_string(self.rate,
                                        self.rate.products[0],
                                        self.rate.reactants[0])
         assert jac == "rho*Y[ic13]*lambda_p_c13__n14"
-
 
     def test_function_string(self):
 
