+++ pynucastro/rates/__init__.py
@@ -1,5 +1,3 @@
 """The core routines needed to read and interpret nuclear reaction rates"""
 
 from .rate import Tfactors, Library, Rate, RateFilter, UnsupportedNucleus, Nucleus, list_known_rates
-
-
