+++ pynucastro/rates/rate.py
@@ -9,7 +9,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 import collections
 
-from pynucastro.nucdata import Element, UnidentifiedElement, PeriodicTable
+from pynucastro.nucdata import UnidentifiedElement, PeriodicTable
+
 
 def list_known_rates():
     """ list the rates found in the library """
@@ -27,6 +28,7 @@ def list_known_rates():
                 for r in lib.get_rates():
                     print("                                 : {}".format(r))
 
+
 class Tfactors(object):
     """ precompute temperature factors for speed """
 
@@ -38,6 +40,7 @@ class Tfactors(object):
         self.T913 = self.T9**(1./3.)
         self.T953 = self.T9**(5./3.)
         self.lnT9 = np.log(self.T9)
+
 
 class SingleSet(object):
     """ a set in Reaclib is one piece of a rate, in the form
@@ -57,11 +60,11 @@ class SingleSet(object):
         self._update_label_properties()
 
     def _update_label_properties(self):
-        """ Set label and flags indicating Set is resonant, 
+        """ Set label and flags indicating Set is resonant,
             weak, or reverse. """
-        assert(type(self.labelprops)==str)
+        assert(type(self.labelprops) == str)
         try:
-            assert(len(self.labelprops)==6)
+            assert(len(self.labelprops) == 6)
         except:
             raise
         else:
@@ -96,7 +99,6 @@ class SingleSet(object):
                                  self.a[5]*tf.T953 +
                                  self.a[6]*tf.lnT9)
 
-
     def set_string(self, prefix="set", plus_equal=False):
         """
         return a string containing the python code for this set
@@ -123,9 +125,11 @@ class SingleSet(object):
         string += ")"
         return string
 
+
 class UnsupportedNucleus(BaseException):
     def __init__(self):
         return
+
 
 class Nucleus(object):
     """
@@ -158,7 +162,7 @@ class Nucleus(object):
             self.N = 1
             self.short_spec_name = "n"
             self.spec_name = "neutron"
-            self.pretty = r"\mathrm{{{}}}".format(self.el)            
+            self.pretty = r"\mathrm{{{}}}".format(self.el)
         else:
             e = re.match(r"([a-zA-Z]*)(\d*)", name)
             self.el = e.group(1).title()  # chemical symbol
@@ -188,10 +192,10 @@ class Nucleus(object):
                 raise
             else:
                 self.Z = i.Z
-                assert(type(self.Z)==int)
+                assert(type(self.Z) == int)
                 assert(self.Z >= 0)
                 self.N = self.A - self.Z
-                assert(type(self.N)==int)
+                assert(type(self.N) == int)
                 assert(self.N >= 0)
 
                 # long name
@@ -216,6 +220,7 @@ class Nucleus(object):
         else:
             return self.A < other.A
 
+
 class Library(object):
     """
     A Library is a Rate container that reads a single file
@@ -238,7 +243,7 @@ class Library(object):
             self._rates = None
             if isinstance(rates, Rate):
                 rates = [rates]
-            assert (isinstance(rates, dict) or isinstance(rates, list)), "ERROR: rates in Library constructor must be a Rate object, list of Rate objects, or dictionary of Rate objects keyed by Rate.get_rate_id()"                
+            assert (isinstance(rates, dict) or isinstance(rates, list)), "ERROR: rates in Library constructor must be a Rate object, list of Rate objects, or dictionary of Rate objects keyed by Rate.get_rate_id()"
             if isinstance(rates, dict):
                 self._rates = rates
             elif isinstance(rates, list):
@@ -385,7 +390,7 @@ class Library(object):
         new_rates = self._rates
         for id, r in other._rates.items():
             try:
[…]
