### 2018-03-10 072ce1c
Capitalize species abbreviations in pretty string. Fixes #45.
