### 2018-03-10 8b7166b
Fix bug where network graph x limit was set using nuclei Z instead of N.
