### 2018-03-10 bec1a28
Set y limit of network plots using nuclei Z.
