+++ pynucastro/networks/rate_collection.py
@@ -337,8 +337,10 @@ class RateCollection(object):
             plt.colorbar(pc)
 
         Ns = [n.N for n in node_nuclei]
+        Zs = [n.Z for n in node_nuclei]
 
         plt.xlim(min(Ns)-1, max(Ns)+1)
+        plt.ylim(min(Zs)-1, max(Zs)+1)
         plt.xlabel(r"$N$", fontsize="large")
         plt.ylabel(r"$Z$", fontsize="large")
 
