### 2018-03-10 4683cdd
Add plotting code for partition functions.
