### 2018-03-10 7fcf110
Add fitting code for partition functions.
