### 2018-03-10 9104ece
Added note about partition functions and CUDA Fortran to future work in paper. Added references.
