+++ paper/paper.bib
@@ -209,3 +209,50 @@ archivePrefix = "arXiv",
    adsurl = {http://adsabs.harvard.edu/abs/2010ApJ...715.1221A},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{RaTh2000,
+   author = {{Rauscher}, T. and {Thielemann}, F.-K.},
+   title = "{Astrophysical Reaction Rates From Statistical Model Calculations}",
+   journal = {Atomic Data and Nuclear Data Tables},
+   eprint = {astro-ph/0004059},
+   year = 2000,
+   month = may,
+   volume = 75,
+   pages = {1-351},
+   doi = {10.1006/adnd.2000.0834},
+   adsurl = {http://adsabs.harvard.edu/abs/2000ADNDT..75....1R},
+   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Rauscher2003,
+   author = {{Rauscher}, T.},
+   title = "{Nuclear Partition Functions at Temperatures Exceeding 10$^{10}$ K}",
+   journal = {\apjs},
+   eprint = {astro-ph/0304047},
+   keywords = {Nuclear Reactions, Nucleosynthesis, Abundances},
+   year = 2003,
+   month = aug,
+   volume = 147,
+   pages = {403-408},
+   doi = {10.1086/375733},
+   adsurl = {http://adsabs.harvard.edu/abs/2003ApJS..147..403R},
+   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{SkyNet.2017,
+   author = {{Lippuner}, J. and {Roberts}, L.~F.},
+   title = "{SkyNet: A Modular Nuclear Reaction Network Library}",
+   journal = {\apjs},
+   archivePrefix = "arXiv",
+   eprint = {1706.06198},
+   primaryClass = "astro-ph.HE",
+   keywords = {methods: numerical, nuclear reactions, nucleosynthesis, abundances },
+   year = 2017,
+   month = dec,
+   volume = 233,
+   eid = {18},
+   pages = {18},
+   doi = {10.3847/1538-4365/aa94cb},
+   adsurl = {http://adsabs.harvard.edu/abs/2017ApJS..233...18L},
+   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
