### 2018-03-10 6417731
Give Zingale 2017 a link. Fix journal names for a couple papers.
