+++ paper/paper.bib
@@ -102,7 +102,7 @@ archivePrefix = "arXiv",
  keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
      year = 2017,
     month = nov,
-   adsurl = {http://adsabs.harvard.edu/abs/2017arXiv171106203Z},
+    url = {http://adsabs.harvard.edu/abs/2017arXiv171106203Z},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -227,7 +227,7 @@ archivePrefix = "arXiv",
 @ARTICLE{Rauscher2003,
    author = {{Rauscher}, T.},
    title = "{Nuclear Partition Functions at Temperatures Exceeding 10$^{10}$ K}",
-   journal = {\apjs},
+   journal = {The Astrophysical Journal Supplement Series},
    eprint = {astro-ph/0304047},
    keywords = {Nuclear Reactions, Nucleosynthesis, Abundances},
    year = 2003,
@@ -242,7 +242,7 @@ archivePrefix = "arXiv",
 @ARTICLE{SkyNet.2017,
    author = {{Lippuner}, J. and {Roberts}, L.~F.},
    title = "{SkyNet: A Modular Nuclear Reaction Network Library}",
-   journal = {\apjs},
+   journal = {The Astrophysical Journal Supplement Series},
    archivePrefix = "arXiv",
    eprint = {1706.06198},
    primaryClass = "astro-ph.HE",
