+++ docs/source/intro.rst
@@ -1,7 +1,7 @@
 Overview of pynucastro
 ==========================
 
-pynucastro is a set of python interfaces nuclear reaction rates
+pynucastro is a set of python interfaces to nuclear reaction rates
 (currently via the JINA reaclib database).  It is meant for both
 interactive exploration of rates (through Jupyter notebooks) and to
 create reaction networks for use in simulation codes.
