### 2018-03-10 6fb51fa
Add note about missing nuclear partition functions to docs.
