+++ docs/source/index.rst
@@ -6,6 +6,15 @@
 pynucastro
 ==============
 
+.. note::
+
+   pynucastro does not yet calculate nuclear partition functions to
+   correct reverse rates in the Reaclib library. Work is ongoing to
+   implement this functionality. We recommend you consider what
+   problem you wish to study using pynucastro to determine whether
+   reverse rates and partition function corrections are significant at
+   the temperatures of interest.
+
 pynucastro is a set of python interfaces to nuclear reaction rate
 databases (including the JINA reaclib nuclear reaction rate database).
 It is meant for both interactive exploration of rates (through Jupyter
