### 2018-03-10 #588 @dwillcox
Thanks @jschwab for the note about the url for the Zingale 2017 paper, which should get a DOI at some point when it is officially published. I've made the link clickable using the `url` bibtex tag as you suggested.

Thank you as well for fixing the typos you've found!
