### 2018-03-10 3be8c3c
Add note about partition functions to readme.
