### 2018-03-10 bd668ee
try deploying docs via Travis.
