+++ .travis.yml
@@ -8,9 +8,9 @@ python:
 before_install:
   - export PATH=$(echo $PATH | tr ':' "\n" | sed '/\/opt\/python/d' | tr "\n" ":" | sed "s|::|:|g")
 
-# addons:
-#   apt_packages:
-#     - pandoc
+addons:
+  apt_packages:
+    - pandoc
 
 install:
   - pip install .
@@ -22,15 +22,15 @@ before_script:
 script:
   - pytest --cov=pynucastro -v
   - pytest --nbval examples -v
-#   - travis-sphinx build
+  - travis-sphinx build
 
-# after_success:
-#   - travis-sphinx deploy
+after_success:
+  - travis-sphinx deploy
 
-# deploy:
-#   provider: pages
-#   skip-cleanup: true
-#   github-token: $GITHUB_TOKEN
-#   keep-history: true
-#   on:
-#     branch: master
+deploy:
+  provider: pages
+  skip-cleanup: true
+  github-token: $GITHUB_TOKEN
+  keep-history: true
+  on:
+    branch: master
