### 2018-03-10 80f1957
test notebooks in the same command as the python scripts.
