+++ .travis.yml
@@ -20,8 +20,7 @@ before_script:
   - export PYTHONPATH=$PYTHONPATH:$(pwd)
 
 script:
-  - pytest --cov=pynucastro -v
-  - pytest --nbval examples -v
+  - pytest -v --cov=. --nbval --ignore=docs
   - travis-sphinx build
 
 after_success:
