### 2018-03-10 #588 @jschwab
@katyhuff @dwillcox @zingale I'm pleased to recommend that the pynucastro paper be accepted in JOSS.  Thanks to the authors for their rapid and helpful communications during the review process.

I checked off the version checkbox, but given the tweaks that happened during review, there should probably be a new minor version released (e.g. v1.1.0) that will correspond to its accepted state.
