### 2018-03-10 302f762
build/deploy docs only with python 3.6 on travis.
