+++ .travis.yml
@@ -21,15 +21,19 @@ before_script:
 
 script:
   - pytest -v --cov=. --nbval --ignore=docs
-#   - travis-sphinx build
+  - if [[ "$TRAVIS_PYTHON_VERSION" == "3.6" ]]; then
+      travis-sphinx build;
+    fi
 
-# after_success:
-#   - travis-sphinx deploy
+after_success:
+  - if [[ "$TRAVIS_PYTHON_VERSION" == "3.6" ]]; then
+      travis-sphinx deploy;
+    fi
 
-# deploy:
-#   provider: pages
-#   skip-cleanup: true
-#   github-token: $GITHUB_TOKEN
-#   keep-history: true
-#   on:
-#     branch: master
+deploy:
+  provider: pages
+  skip-cleanup: true
+  github-token: $GITHUB_TOKEN
+  keep-history: true
+  on:
+    branch: master
