### 2018-03-10 013964f
Require Sphinx < 1.7.1 to avoid incompatibility with travis-sphinx introduced in version 1.7.1. See https://github.com/Syntaf/travis-sphinx/issues/34 for details.
