+++ requirements.txt
@@ -7,6 +7,7 @@ jupyter>=1.0.0
 ipywidgets>=7.1.2
 travis-sphinx
 sphinx_rtd_theme
+Sphinx<1.7.1
 nbsphinx>=0.3.1
 pytest>=3.2.1
 nbval>=0.9.0
