### 2018-03-10 c95abcc
Set user name and email for git with Travis-CI.
