+++ .travis.yml
@@ -41,4 +41,3 @@ deploy:
   keep-history: true
   on:
     branch: master
-
