+++ .travis.yml
@@ -29,10 +29,10 @@ script:
       travis-sphinx build;
     fi
 
-after_success:
-  - if [[ "$TRAVIS_PYTHON_VERSION" == "3.6" ]]; then
-      travis-sphinx deploy;
-    fi
+# after_success:
+#   - if [[ "$TRAVIS_PYTHON_VERSION" == "3.6" ]]; then
+#       travis-sphinx deploy;
+#     fi
 
 deploy:
   provider: pages
