### 2018-03-11 1606c38
Quote passphrase in gpg command to escape special characters.
