+++ _sources/pynucastro.nucdata.rst.txt
@@ -1,5 +1,5 @@
-pynucastro\.nucdata package
-===========================
+pynucastro.nucdata package
+==========================
 
 .. automodule:: pynucastro.nucdata
     :members:
@@ -9,18 +9,26 @@ pynucastro\.nucdata package
 Submodules
 ----------
 
-pynucastro\.nucdata\.binding\_nuclide module
---------------------------------------------
+pynucastro.nucdata.binding\_nuclide module
+------------------------------------------
 
 .. automodule:: pynucastro.nucdata.binding_nuclide
     :members:
     :undoc-members:
     :show-inheritance:
 
-pynucastro\.nucdata\.binding\_table module
-------------------------------------------
+pynucastro.nucdata.binding\_table module
+----------------------------------------
 
 .. automodule:: pynucastro.nucdata.binding_table
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pynucastro.nucdata.elements module
+----------------------------------
+
+.. automodule:: pynucastro.nucdata.elements
     :members:
     :undoc-members:
     :show-inheritance:
