+++ _sources/library-filtering-examples.ipynb.txt
@@ -33,7 +33,9 @@
    "cell_type": "code",
    "execution_count": 1,
    "metadata": {
-    "collapsed": true
+    "tags": [
+     "nbval-skip"
+    ]
    },
    "outputs": [],
    "source": [
@@ -45,7 +47,9 @@
    "cell_type": "code",
    "execution_count": 2,
    "metadata": {
-    "collapsed": true
+    "tags": [
+     "nbval-skip"
+    ]
    },
    "outputs": [],
    "source": [
@@ -56,7 +60,9 @@
    "cell_type": "code",
    "execution_count": 3,
    "metadata": {
-    "collapsed": true
+    "tags": [
+     "nbval-skip"
+    ]
    },
    "outputs": [],
    "source": [
@@ -80,7 +86,9 @@
    "cell_type": "code",
    "execution_count": 4,
    "metadata": {
-    "collapsed": true
+    "tags": [
+     "nbval-skip"
+    ]
    },
    "outputs": [],
    "source": [
@@ -101,31 +109,35 @@
   {
    "cell_type": "code",
    "execution_count": 5,
-   "metadata": {},
+   "metadata": {
+    "tags": [
+     "nbval-skip"
+    ]
+   },
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "c12 --> n + c11    (c12 --> n + c11_bb92_reaclib_reverse)\n",
-      "c12 --> p + b11    (c12 --> p + b11_nw00_reaclib_reverse)\n",
-      "c12 --> he4 + he4 + he4    (c12 --> he4 + he4 + he4_fy05_reaclib_reverse)\n",
-      "c12 + n --> c13    (c12 + n --> c13_ka02_reaclib)\n",
-      "c12 + p --> n13    (c12 + p --> n13_ls09_reaclib)\n",
-      "c12 + he4 --> o16    (c12 + he4 --> o16_nac2_reaclib)\n",
-      "c12 + n --> p + b12    (c12 + n --> p + b12_wag_reaclib_reverse)\n",
-      "c12 + n --> he4 + be9    (c12 + n --> he4 + be9_cf88_reaclib_reverse)\n",
-      "c12 + he4 --> n + o15    (c12 + he4 --> n + o15_cf88_reaclib_reverse)\n",
[…]
