### 2018-03-11 19ad44f
Revise requirements so we no longer need travis-sphinx or Sphinx < 1.7.1. Also relax requirement for matplotlib==2.1.0.
