+++ requirements.txt
@@ -2,12 +2,11 @@ networkx>=2.1
 scipy>=1.0.0
 numpy>=1.13.3
 sympy>=1.1.1
-matplotlib==2.1.0
+matplotlib>=2.1.0
 jupyter>=1.0.0
 ipywidgets>=7.1.2
-travis-sphinx
 sphinx_rtd_theme
-Sphinx<1.7.1
+Sphinx
 nbsphinx>=0.3.1
 pytest>=3.2.1
 nbval>=0.9.0
