### 2018-03-12 #588 @kyleniemeyer
@dwillcox @zingale a few minor comments:

- It might be helpful to add some instructions on installing the package—it could be as simple as adding a brief section in the README (even if they are pretty standard instructions)
- There are a few references in the paper with URLs that should be replaced with DOIs: Cyburt et al. 2010: 10.1088/0067-0049/189/1/240, Huang et al. 2017: 10.1088/1674-1137/41/3/030002, Nonaka et al. 2010: 10.1088/0067-0049/188/2/358; Suzuki et al. 2016: 10.3847/0004-637X/817/2/163; Wang et al. 2017: 10.1088/1674-1137/41/3/030003
