### 2018-03-12 #588 @kyleniemeyer
Oh, one other comment: it might be helpful to have a link from the docs page back to the repository; I couldn't find one.
