### 2018-03-12 7e18f6f
use find_packages() to get subpackages recursively
