+++ setup.py
@@ -8,6 +8,7 @@ setup(name='pynucastro',
       author_email='michael.zingale@stonybrook.edu',
       license='BSD',
       packages=find_packages(),
+      package_data={"pynucastro": ["library/*", "library/tabular/*"]},
       install_requires=['networkx', 'numpy',
                         'sympy', 'scipy', 'matplotlib',
                         'ipywidgets'],
