### 2018-03-12 27bb810
add install instructions
