### 2018-03-12 #588 @zingale
I've added a link to the project github on the docs index page.

I also added install instructions (and fixed an issue in the `setup.py`) and confirmed that the example notebook runs with the library installed via `setup.py`

I'll work on the references.
