### 2018-03-12 936c2d4
add DOIs
