+++ paper/paper.bib
@@ -9,7 +9,7 @@
   volume={189},
   number={1},
   pages={240},
-  url={http://stacks.iop.org/0067-0049/189/i=1/a=240},
+  url={https://doi.org/10.1088/0067-0049/189/1/240},
   year={2010}
 }
 
@@ -20,7 +20,7 @@
   volume={817},
   number={2},
   pages={163},
-  url={http://stacks.iop.org/0004-637X/817/i=2/a=163},
+  url={https://doi.org/10.3847/0004-637x/817/2/163},
   year={2016}
 }
 
@@ -52,7 +52,7 @@
   volume={41},
   number={3},
   pages={030002},
-  url={http://stacks.iop.org/1674-1137/41/i=3/a=030002},
+  url={https://doi.org/10.1088/1674-1137/41/3/030002},
   year={2017}
 }
 
@@ -63,7 +63,7 @@
   volume={41},
   number={3},
   pages={030003},
-  url={http://stacks.iop.org/1674-1137/41/i=3/a=030003},
+  url={https://doi.org/10.1088/1674-1137/41/3/030003},
   year={2017}
 }
 
@@ -187,7 +187,8 @@ eprint = {https://doi.org/10.1137/0910062}
      year = {2010},
    volume = {188},
     pages = {358-383},
-     note = {paper V}
+     note = {paper V},
+    url = {https://doi.org/10.1088/0067-0049/188/2/358}
 }
 
 @ARTICLE{castro,
