### 2018-03-12 #588 @zingale
okay, I think all the DOIs that are available are in place now.
