### 2018-03-12 #588 @jschwab
@zingale The install instructions in https://github.com/pynucastro/pynucastro/commit/27bb810131ce4dad60ebb192150f7d01de6569f5 trail off.  "This will put the..."
