### 2018-03-13 #588 @dwillcox
Hi @katyhuff @jschwab @kyleniemeyer,

Thank you for all your helpful review suggestions that have significantly improved the usability and accessibility of pynucastro.

Today I pushed a commit adding a brief section to the documentation describing a previously undocumented feature of pynucastro: [Specifying Desired Nuclei to Construct a Network](https://pynucastro.github.io/pynucastro/library-examples-nuclei.html)

I have just tagged two commits as follows on github:

- v1.0.0: the version we submitted to JOSS
- v1.1.0: the current version after incorporating reviewer comments.

We are ready for whatever the next steps are, please let us know.

Thanks again!
Don
