### 2018-03-13 #588 @jschwab
> Today I pushed a commit adding a brief section to the documentation describing a previously undocumented feature of pynucastro: Specifying Desired Nuclei to Construct a Network

Just looked that bit over.  That's a nice feature.
