### 2018-03-18 #588 @katyhuff
Wonderful! Thanks @dwillcox  @zingale @jschwab & @kyleniemeyer !

 @dwillcox  & @zingale : Could you make an archive of the current version of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
