### 2018-03-18 #588 @dwillcox
Also, I had to make a new release on github in order for Zenodo to archive it, so this is v1.1.1.
