### 2018-03-18 1e5e9d4
Add Zenodo DOI badge.
