### 2018-03-19 #588 @dwillcox
Thanks again @katyhuff @jschwab @kyleniemeyer -- it's been a pleasure interacting with you as we prepared pynucastro for publication via JOSS.

This was my first time using JOSS and this has been a very constructive experience!

-Don
