### 2018-03-19 #588 @dwillcox
Quick question @katyhuff -- since we retain copyright to the JOSS paper and it's licensed as CC-BY, can we submit the JOSS paper to the arXiv preprint server without modification?
